build.sh: Add parameter command and modified kernel config

+++ build.sh
@@ -1,22 +1,45 @@
 #!/bin/bash
 usage()
 {
-    echo "USAGE: [-o] [-u] [-v VERSION_NAME]"
-    echo "No ARGS means use default build option"
-    echo "WHERE: -o = generate ota package       "
-    echo "       -u = generate update.img        "
-    echo "       -v = set build version name for output image folder"
+   echo "USAGE: [-U] [-K] [-A] [-p] [-o] [-u] [-v VERSION_NAME]  "
+    echo "No ARGS means use default build option                  "
+    echo "WHERE: -U = build uboot                                 "
+    echo "       -K = build kernel                                "
+    echo "       -A = build android                               "
+    echo "       -p = will build packaging in IMAGE      "
+    echo "       -o = build OTA package                           "
+    echo "       -u = build update.img                            "
+    echo "       -v = build android with 'user' or 'userdebug'    "
     exit 1
 }
-
+BUILD_UBOOT=false
+BUILD_KERNEL=false
+BUILD_ANDROID=false
 BUILD_UPDATE_IMG=false
 BUILD_OTA=false
-BUILD_VERSION="IMAGES"
+BUILD_PACKING=false
+
 
 # check pass argument
-while getopts "ouv:" arg
+while getopts "UKApouv:" arg
 do
     case $arg in
+        U)
+            echo "will build u-boot"
+            BUILD_UBOOT=true
+            ;;
+        K)
+            echo "will build kernel"
+            BUILD_KERNEL=true
+            ;;
+        A)
+            echo "will build android"
+            BUILD_ANDROID=true
+            ;;
+        p)
+            echo "will build packaging in IMAGE"
+            BUILD_PACKING=true
+            ;;
         o)
             echo "will build ota package"
             BUILD_OTA=true
@@ -26,14 +49,15 @@ do
             BUILD_UPDATE_IMG=true
             ;;
         v)
-            BUILD_VERSION=$OPTARG
+            BUILD_VARIANT=$OPTARG
             ;;
         ?)
             usage ;;
     esac
 done
 
-source build/envsetup.sh >/dev/null && setpaths
+
+source build/envsetup.sh >/dev/null
 TARGET_PRODUCT=`get_build_var TARGET_PRODUCT`
 
 #set jdk version
@@ -43,9 +67,9 @@ export CLASSPATH=.:$JAVA_HOME/lib:$JAVA_HOME/lib/tools.jar
 # source environment and chose target product
 DEVICE=`get_build_var TARGET_PRODUCT`
 BUILD_VARIANT=`get_build_var TARGET_BUILD_VARIANT`
-UBOOT_DEFCONFIG=rk3399_defconfig
+UBOOT_DEFCONFIG=rk3399
 KERNEL_DEFCONFIG=rockchip_defconfig
-KERNEL_DTS=rk3399-sapphire-excavator-edp
+KERNEL_DTS=rk3399-sapphire-excavator-edp-avb
 PACK_TOOL_DIR=RKTools/linux/Linux_Pack_Firmware
 IMAGE_PATH=rockdev/Image-$TARGET_PRODUCT
 export PROJECT_TOP=`gettop`
@@ -60,36 +84,40 @@ export STUB_PATH=$PROJECT_TOP/$STUB_PATH
 export STUB_PATCH_PATH=$STUB_PATH/PATCHES
 
 # build uboot
+if [ "$BUILD_UBOOT" = true ] ; then
 echo "start build uboot"
-cd u-boot && make distclean && make $UBOOT_DEFCONFIG && ./mkv8.sh && cd -
+cd u-boot && ./make.sh $UBOOT_DEFCONFIG && cd -
 if [ $? -eq 0 ]; then
     echo "Build uboot ok!"
 else
     echo "Build uboot failed!"
     exit 1
 fi
-
+fi
 
 # build kernel
+if [ "$BUILD_KERNEL" = true ] ; then
 echo "Start build kernel"
-cd kernel && make clean &&make ARCH=arm64 $KERNEL_DEFCONFIG && make ARCH=arm64 $KERNEL_DTS.img -j12 && cd -
+cd kernel && make ARCH=arm64 $KERNEL_DEFCONFIG && make ARCH=arm64 $KERNEL_DTS.img -j24 && cd -
 if [ $? -eq 0 ]; then
     echo "Build kernel ok!"
 else
     echo "Build kernel failed!"
     exit 1
 fi
-
+fi
 
 # build android
+if [ "$BUILD_ANDROID" = true ] ; then
 echo "start build android"
 make installclean
-make -j12
+make -j24
 if [ $? -eq 0 ]; then
     echo "Build android ok!"
 else
     echo "Build android failed!"
     exit 1
+fi
 fi
 
 # mkimage.sh
@@ -113,27 +141,31 @@ if [ "$BUILD_OTA" = true ] ; then
 fi
 
 
-mkdir -p $PACK_TOOL_DIR/rockdev/Image/
-cp -f $IMAGE_PATH/* $PACK_TOOL_DIR/rockdev/Image/
+if [ "$BUILD_UPDATE_IMG" = true ] ; then
+    mkdir -p $PACK_TOOL_DIR/rockdev/Image/
+    cp -f $IMAGE_PATH/* $PACK_TOOL_DIR/rockdev/Image/
 
-echo "Make update.img"
-cd $PACK_TOOL_DIR/rockdev && ./mkupdate.sh
-if [ $? -eq 0 ]; then
-    echo "Make update image ok!"
-else
-    echo "Make update image failed!"
-    exit 1
+    echo "Make update.img"
+    cd $PACK_TOOL_DIR/rockdev && ./mkupdate.sh
+    if [ $? -eq 0 ]; then
+        echo "Make update image ok!"
+    else
+        echo "Make update image failed!"
+        exit 1
+    fi
+    cd -
+    mv $PACK_TOOL_DIR/rockdev/update.img $IMAGE_PATH/
+    rm $PACK_TOOL_DIR/rockdev/Image -rf
 fi
-cd -
 
-mv $PACK_TOOL_DIR/rockdev/update.img $IMAGE_PATH/
-rm $PACK_TOOL_DIR/rockdev/Image -rf
+if [ "$BUILD_PACKING" = true ] ; then
+echo "make and copy packaging in IMAGE "
 
 mkdir -p $STUB_PATH
 
 #Generate patches
 
-.repo/repo/repo forall  -c '[ "$REPO_REMOTE" = "rk" ] && { REMOTE_DIFF=`git log $REPO_REMOTE/$REPO_RREV..HEAD`; LOCAL_DIFF=`git diff`; [ -n "$REMOTE_DIFF" ] && { mkdir -p $STUB_PATCH_PATH/$REPO_PATH/; git format-patch $REPO_REMOTE/$REPO_RREV..HEAD -o $STUB_PATCH_PATH/$REPO_PATH; git merge-base HEAD $REPO_REMOTE/$REPO_RREV | xargs git show -s > $STUB_PATCH_PATH/$REPO_PATH/git-merge-base.txt; } || :; [ -n "$LOCAL_DIFF" ] && { mkdir -p $STUB_PATCH_PATH/$REPO_PATH/; git reset HEAD ./; git diff > $STUB_PATCH_PATH/$REPO_PATH/local_diff.patch; } || :; }'
+.repo/repo/repo forall  -c "$PROJECT_TOP/device/rockchip/common/gen_patches_body.sh"
 
 #Copy stubs
 cp commit_id.xml $STUB_PATH/manifest_${DATE}.xml
@@ -149,3 +181,4 @@ cp build.sh $STUB_PATH/build.sh
 echo "UBOOT:  defconfig: $UBOOT_DEFCONFIG" >> $STUB_PATH/build_cmd_info
 echo "KERNEL: defconfig: $KERNEL_DEFCONFIG, dts: $KERNEL_DTS" >> $STUB_PATH/build_cmd_info
 echo "ANDROID:$DEVICE-$BUILD_VARIANT" >> $STUB_PATH/build_cmd_info
+fi
